requirements.txt
python-telegram-bot==20.7
python-dotenv==0.19.2
openai>=1.0.0
aiohttp==3.8.5